TST: collect .insert tests (#32909)
